refactor(blast): one evidence helper, used by the page and the comment
The comment's collapsed symbol list cited a span; it now quotes the line that
reaches the diff, from the same helper the hosted panel uses — a comment that
disagreed with the page it links to would be worse than one that said less. The
quote stays below the fold, so the comment is still five circles and a table.
